refactor(rl): structurally separate V3/V4 into independent subdirectories
Move V3-specific files (tokenization, model, cache, datasets) into
pymahjong/rl/v3/ and V4-specific files into pymahjong/rl/v4/, with
thin backward-compat shim files at the rl/ level.

Extract TransformerConfig into rl/common/config.py to break the V4→V3
model dependency. Extend EncodingStrategy protocol with create_dataset,
obs_to_tensor, forward_from_batch, and evaluate_actions_from_batch to
eliminate all encoding-specific if-else branches from bc.py and ppo.py.
