+++ pymahjong/rl/common/__init__.py
@@ -1,0 +1,1 @@
+"""Shared components used across encoding versions."""
