+++ pymahjong/rl/encodings/v4.py
@@ -2,7 +2,9 @@
 
 from __future__ import annotations
 
-from typing import Any, Dict, List, Optional
+import os
+import glob as _glob
+from typing import Any, Dict
 
 import numpy as np
 
@@ -19,11 +21,7 @@ class V4Strategy:
     # -- Observation -----------------------------------------------------------
 
     def encode_observation(self, table, current_player: int, **kwargs) -> Dict[str, Any]:
-        # V4 encoding is event-driven; a single-table snapshot requires
-        # replaying the game log.  Use StreamingPaipuDatasetV4 for data
-        # generation.  For online env use, we produce a padded mask from
-        # the current table state.
-        from ..tokenization_v4 import _engine_action_mask
+        from ..v4.tokenization import _engine_action_mask
 
         mask = _engine_action_mask(table, current_player)
         return {
@@ -49,9 +47,8 @@ class V4Strategy:
     # -- Model -----------------------------------------------------------------
 
     def create_model(self, **kwargs) -> Any:
-        from ..model_v4 import EventStreamTransformer
-
-        from ..model import TransformerConfig
+        from ..common.config import TransformerConfig
+        from ..v4.model import EventStreamTransformer
 
         cfg = kwargs.get("transformer_config") or TransformerConfig()
         return EventStreamTransformer(config=cfg, event_dim=self.EVENT_DIM)
@@ -59,26 +56,85 @@ class V4Strategy:
     # -- Collation -------------------------------------------------------------
 
     def collate_fn(self, batch: list) -> Dict[str, Any]:
-        from ..cached_dataset_v4 import cached_event_collate
+        from ..v4.cached_dataset import cached_event_collate
 
         return cached_event_collate(batch)
 
     # -- Cache / Dataset factories ---------------------------------------------
 
     def create_shard_writer(self, shard_dir: str, **kwargs):
-        from ..cache_v4 import V4ShardWriter
+        from ..v4.cache import V4ShardWriter
 
         return V4ShardWriter(shard_dir)
 
     def create_cached_dataset(self, cache_dir: str, **kwargs):
-        from ..cached_dataset_v4 import CachedEventDataset
+        from ..v4.cached_dataset import CachedEventDataset
 
         return CachedEventDataset(cache_dir, **kwargs)
 
     def create_streaming_dataset(self, paths, **kwargs):
-        from ..tokenization_v4 import StreamingPaipuDatasetV4
+        from ..v4.tokenization import StreamingPaipuDatasetV4
 
         return StreamingPaipuDatasetV4(paipu_paths=paths, **kwargs)
+
+    # -- Training integration --------------------------------------------------
+
+    def create_dataset(self, mode: str, config=None, **kwargs):
+        """Create a training dataset for V4.
+
+        Args:
+            mode: ``"cached"`` or ``"streaming"``.
+            config: ``BCConfig`` (or compatible) with ``cache_dir``,
+                ``paipu_dir``, etc.
+        """
+        if mode == "cached":
+            from ..v4.cached_dataset import CachedEventDataset
+            return CachedEventDataset(config.cache_dir)
+        elif mode == "streaming":
+            from ..v4.tokenization import StreamingPaipuDatasetV4
+            paipu_dir = getattr(config, "paipu_dir", None)
+            if not paipu_dir:
+                from pymahjong.config import get_config
+                paipu_dir = get_config().paipu_xml_path
+            paths = sorted(
+                p for p in _glob.glob(os.path.join(paipu_dir, "**", "*"), recursive=True)
+                if os.path.isfile(p) and (p.endswith(".xml") or p.endswith(".txt"))
+            )
+            return StreamingPaipuDatasetV4(
+                paipu_paths=paths,
+                prefetch_n=getattr(config, "paipu_prefetch", 4),
+            )
+        else:
+            raise ValueError(f"V4 does not support dataset mode: {mode!r}")
+
+    def obs_to_tensor(self, obs: dict, device):
+        """Convert a single V4 observation dict to model-ready tensors."""
+        import torch
+        feat = torch.as_tensor(obs["features"], device=device).unsqueeze(0)
+        if not feat.is_floating_point():
+            feat = feat.float()
+        return (
+            feat,
+            torch.as_tensor(obs["attention_mask"], device=device, dtype=torch.bool).unsqueeze(0),
+            torch.as_tensor(obs["action_mask"], device=device, dtype=torch.bool).unsqueeze(0),
+        )
+
+    def forward_from_batch(self, model, batch: dict):
+        """Dispatch a V4 batch through the model → (logits, value)."""
+        return model(
+            batch["features"],
+            batch["attention_mask"],
+            batch["action_mask"],
+        )
+
+    def evaluate_actions_from_batch(self, model, batch: dict, actions):
+        """Dispatch a V4 batch through model.evaluate_actions."""
+        return model.evaluate_actions(
+            batch["features"],
+            batch["attention_mask"],
+            batch["action_mask"],
+            actions,
+        )
 
 
 register(EncodingVersion.V4, V4Strategy())
