+++ pymahjong/rl/v3/__init__.py
@@ -1,0 +1,1 @@
+"""V3 token-based encoding strategy."""
