+++ pymahjong/rl/v4/__init__.py
@@ -1,0 +1,1 @@
+"""V4 event-stream encoding strategy."""
