+++ pymahjong/rl/v4/collate.py
@@ -1,0 +1,5 @@
+"""V4 collation functions for batching observations."""
+
+from .cached_dataset import cached_event_collate  # noqa: F401
+
+__all__ = ["cached_event_collate"]
